# api/openapi.yaml
openapi: 3.0.0
info:
  title: Wallet Service API
  version: 1.0.0
servers:
  - url: http://localhost:8080

paths:
  /health:
    get:
      operationId: HealthCheck
      summary: Проверка работоспособности сервиса
      responses:
        '200':
          description: Сервис работает нормально
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"

  /api/v1/wallet:
    post:
      operationId: ProcessWalletOperation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletOperationRequest'
      responses:
        '204':
          description: Операция успешно выполнена
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Кошелёк не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Недостаточно средств
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/wallets:
    post:
      operationId: CreateWallet
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '201':
          description: Кошелёк успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalanceResponse'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Кошелёк уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/wallets/{walletId}:
    get:
      operationId: GetWalletBalance
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Баланс кошелька
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalanceResponse'
        '400':
          description: Некорректный UUID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Кошелёк не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    WalletOperationRequest:
      type: object
      required: [walletId, operationType, amount]
      properties:
        walletId:
          type: string
          format: uuid
        operationType:
          type: string
          enum: [DEPOSIT, WITHDRAW]
        amount:
          type: integer
          format: int64
          minimum: 1


    WalletBalanceResponse:
      type: object
      properties:
        walletId:
          type: string
          format: uuid
        balance:
          type: integer
          format: int64

    Error:
      type: object
      properties:
        message:
          type: string