# internal/generated/api.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package generated

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for WalletOperationRequestOperationType.
const (
	DEPOSIT  WalletOperationRequestOperationType = "DEPOSIT"
	WITHDRAW WalletOperationRequestOperationType = "WITHDRAW"
)

// Error defines model for Error.
type Error struct {
	Message *string `json:"message,omitempty"`
}

// WalletBalanceResponse defines model for WalletBalanceResponse.
type WalletBalanceResponse struct {
	Balance  *int64              `json:"balance,omitempty"`
	WalletId *openapi_types.UUID `json:"walletId,omitempty"`
}

// WalletOperationRequest defines model for WalletOperationRequest.
type WalletOperationRequest struct {
	Amount        int64                               `json:"amount"`
	OperationType WalletOperationRequestOperationType `json:"operationType"`
	WalletId      openapi_types.UUID                  `json:"walletId"`
}

// WalletOperationRequestOperationType defines model for WalletOperationRequest.OperationType.
type WalletOperationRequestOperationType string

// CreateWalletJSONBody defines parameters for CreateWallet.
type CreateWalletJSONBody = map[string]interface{}

// ProcessWalletOperationJSONRequestBody defines body for ProcessWalletOperation for application/json ContentType.
type ProcessWalletOperationJSONRequestBody = WalletOperationRequest

// CreateWalletJSONRequestBody defines body for CreateWallet for application/json ContentType.
type CreateWalletJSONRequestBody = CreateWalletJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /api/v1/wallet)
	ProcessWalletOperation(w http.ResponseWriter, r *http.Request)

	// (POST /api/v1/wallets)
	CreateWallet(w http.ResponseWriter, r *http.Request)

	// (GET /api/v1/wallets/{walletId})
	GetWalletBalance(w http.ResponseWriter, r *http.Request, walletId openapi_types.UUID)
	// Проверка работоспособности сервиса
	// (GET /health)
	HealthCheck(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// (POST /api/v1/wallet)
func (_ Unimplemented) ProcessWalletOperation(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /api/v1/wallets)
func (_ Unimplemented) CreateWallet(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api/v1/wallets/{walletId})
func (_ Unimplemented) GetWalletBalance(w http.ResponseWriter, r *http.Request, walletId openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Проверка работоспособности сервиса
// (GET /health)
func (_ Unimplemented) HealthCheck(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ProcessWalletOperation operation middleware
func (siw *ServerInterfaceWrapper) ProcessWalletOperation(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ProcessWalletOperation(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateWallet operation middleware
func (siw *ServerInterfaceWrapper) CreateWallet(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateWallet(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetWalletBalance operation middleware
func (siw *ServerInterfaceWrapper) GetWalletBalance(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "walletId" -------------
	var walletId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "walletId", chi.URLParam(r, "walletId"), &walletId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "walletId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetWalletBalance(w, r, walletId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// HealthCheck operation middleware
func (siw *ServerInterfaceWrapper) HealthCheck(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.HealthCheck(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/v1/wallet", wrapper.ProcessWalletOperation)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/v1/wallets", wrapper.CreateWallet)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/wallets/{walletId}", wrapper.GetWalletBalance)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/health", wrapper.HealthCheck)
	})

	return r
}